Failing feature for filtering of query params
Although sensitive data included in the query string of URL is filtered out of the
params information sent to HoneyBadger it is still left in the request sent to
HoneyBadger and is then displayed in the HoneyBadger interface in the URL and
environment variables.

See https://github.com/honeybadger-io/honeybadger-ruby/issues/40 for more details.